Accounting AI_v2 better/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pypdf==4.2.0
slowapi==0.1.9
